+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: remove trailing slash in `scim_url`
 - Fix: send PATCH request for change of displayName of Group (e.g. the name of the Role)
+- Fix: accept invitation in same window as where the invitation was send, now works
 
 ## 1.0.2